Fix for NCL-MED-03 – Cross-Frame Scripting (XFS)
NCL-MED-03 – Cross-Frame Scripting (XFS)
It is advised that the web application should set the appropriate value
for X-Frame-Options security header on the server to restrict the
website from being embedded in untrusted third-party websites. It is
suggested to set the security header value to sameorigin (only allow
current domain to frame itself) or deny (no frames allowed).